requirements.txt
asgiref==3.6.0
Django==3.2.16
django-cleanup==7.0.0
django-debug-toolbar==4.0.0
django-filter==23.1
djangorestframework==3.14.0
Pillow==9.5.0
pytz==2022.7.1
routers==0.10.1
sqlparse==0.4.3
